# schemas/openapi-commercial-v1.yaml
openapi: "3.0.3"
info:
  title: AgentSkeptic commercial license API
  version: "1.0.0"
  contact:
    url: https://agentskeptic.com
  x-agentskeptic-distribution:
    repository: https://github.com/jwekavanagh/agentskeptic
    npmPackage: https://www.npmjs.com/package/agentskeptic
    openApi: https://agentskeptic.com/openapi-commercial-v1.yaml
  description: "Read-only checks at verify time—not color.\n\nMachine-readable contract for license preflight used by the published npm CLI.\nBase URL is your deployed app origin (same as NEXT_PUBLIC_APP_URL).\n\nEvery path in this document returns an `x-request-id` response header on all status codes (echo a valid client `x-request-id` when supplied; otherwise server-generated). Non-2xx JSON bodies follow RFC 7807-style Problem Details (`type`, `title`, `status`, `detail`, optional `code`, `instance`) unless noted; `POST /api/v1/usage/reserve` denials additionally include legacy `allowed`, `code`, `message`, and optional `upgrade_url` for backward compatibility.\n"
...
servers:
  - url: https://agentskeptic.com
    description: Replace with your deployment origin
paths:
  /api/v1/usage/reserve:
    post:
      operationId: reserveUsage
      summary: Reserve one verification or enforcement run (license preflight)
      description: >
        Called by the commercial CLI before contract verify, quick verify, or enforce.
        Requires Bearer API key. Body run_id must be idempotent per key.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveRequest"
      responses:
        "200":
          description: Allowed; quota consumed or duplicate run_id returned without double-charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveAllowed"
        "400":
          description: Bad request (missing/invalid body, skewed issued_at)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
        "401":
          description: Invalid API key or key lifecycle denial (revoked, disabled, expired)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
        "403":
          description: Quota exceeded or entitlement denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
        "503":
          description: Server error during reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
...
  /api/v1/funnel/verify-outcome:
    post:
      operationId: postVerifyOutcomeBeacon
      summary: Licensed verify-outcome beacon (idempotent per API key + run_id)
      description: >
        POST after a successful `POST /api/v1/usage/reserve` for the same `run_id`. Success is HTTP 204 with an empty body.
        Requires Bearer API key. Body schemaVersion must be 2.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOutcomeRequestV2"
      responses:
        "204":
          description: Beacon accepted (or duplicate ignored)
        "400":
          description: Invalid JSON or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No reservation for this run_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "410":
          description: Reservation too old
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
    VerifyOutcomeRequestV2:
      type: object
      required:
        - schema_version
        - run_id
        - workflow_id
        - trust_decision
        - reason_codes
        - terminal_status
        - workload_class
        - subcommand
      properties:
        schema_version:
          type: integer
          const: 2
        run_id:
          type: string
          maxLength: 256
        workflow_id:
          type: string
          maxLength: 512
        trust_decision:
          type: string
          enum: [safe, unsafe, unknown]
        reason_codes:
          type: array
          maxItems: 8
          items:
            type: string
        terminal_status:
          type: string
          enum: [complete, inconsistent, incomplete]
        workload_class:
          type: string
          enum: [bundled_examples, non_bundled]
        subcommand:
          type: string
          enum: [batch_verify, quick_verify, verify_integrator_owned]
...
    ReserveRequest:
      type: object
      required:
        - run_id
        - issued_at
      properties:
        run_id:
          type: string
          maxLength: 256
        issued_at:
          type: string
          format: date-time
          description: ISO8601; server rejects if skew from now exceeds 300 seconds
        intent:
          type: string
          enum: [verify, enforce]
          default: verify
          description: >
            enforce requires a paid plan (not Starter) with active subscription; verify is allowed
            for Starter (free included quota) and for paid plans with active subscription; overage
            is permitted past included for paid plans with allowOverage in plan catalog
    ReserveAllowed:
      type: object
      required:
        - allowed
        - plan
        - limit
        - used
        - included_monthly
        - overage_count
      properties:
        allowed:
          type: boolean
          enum: [true]
        plan:
          type: string
          enum: [starter, individual, team, business, enterprise]
        limit:
          type: integer
          description: Included verifications in the current billing month (not enterprise unlimited sentinel)
        used:
          type: integer
          description: Total verifications counted this month for this API key
        included_monthly:
          oneOf:
            - type: integer
            - type: "null"
          description: Same as limit for finite plans; null for enterprise unlimited
        overage_count:
          type: integer
          description: max(0, used - included_monthly) when the plan has a finite cap
    ReserveError:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          required: [allowed, code, message]
          properties:
            allowed:
              type: boolean
              enum: [false]
            code:
              type: string
              description: >
                Examples BAD_REQUEST, INVALID_KEY, QUOTA_EXCEEDED, ENFORCEMENT_REQUIRES_PAID_PLAN,
                VERIFICATION_REQUIRES_SUBSCRIPTION, SUBSCRIPTION_INACTIVE, BILLING_PRICE_UNMAPPED,
                INSUFFICIENT_SCOPE, KEY_EXPIRED, KEY_REVOKED, KEY_DISABLED, SERVER_ERROR
            message:
              type: string
            upgrade_url:
              type: string
              format: uri
              description: Present for some 403 entitlement responses
...